Added ability to grab and drop boxes to the crane
Since there was no prefab to work with, I created my own box to test on. 

The grabbing only works iff the pivot is set to the top of the box, and the rigidbody variable is set to it's own rigidbody. 

This may make it difficult to use it as a prefab, however this can be tested later.